-Fixed metaTiffRead so that it now closes files properly. -Fixed bug in moveSTKFiles2OwnDir. -Modified trackCloseGapsKalmanSparse to put zeros and NaNs in their proper place when converting from sparse to full (before I was taking a shortcut). -Fixed bug in spatialMovAveBG that was causing a mess when images were cropped. -Added new scripts scriptCalcFeatDensity and scriptDetectGeneralBatch
git-svn-id: https://orchestra.med.harvard.edu/svn/lccb/matlab/projects/common/main@3780 0bcb941b-74d1-0310-bb26-e6636707f451

+++ detectionAlgorithms/subResFeatures2D/detectSubResFeatures2D_StandAlone.m
@@ -228,7 +228,7 @@ for iImage = 1 : numImagesInteg
     
     %store raw images in array
     for jImage = 1 : 1 + 2*integWindow;
-        imageRaw(:,:,jImage) = imread([imageDir filenameBase enumString(imageIndx(jImage+iImage-1),:) '.tif']);
+        imageRaw(:,:,jImage) = double(imread([imageDir filenameBase enumString(imageIndx(jImage+iImage-1),:) '.tif']));
     end
     
     %     %display progress
@@ -241,7 +241,7 @@ for iImage = 1 : numImagesInteg
     imageRaw(imageRaw==0) = NaN;
     
     %normalize images
-    imageRaw = double(imageRaw) / (2^bitDepth-1);
+    imageRaw = imageRaw / (2^bitDepth-1);
     
     % %integrate over time
     % while numImagesInteg <= 0
